adding error handlers
adding error handlers for register, login, createpack, aquirepack

+++ src/main/java/at/fhtw/bif3vz/swe/mtcg/if19b101/database/DatabaseOperations.java
@@ -22,14 +22,26 @@ public class DatabaseOperations {
         }
     }
 
-    public static void writeUserToDatabase(TestUser user) {
-        Database daoDb = new Database();
+    public static boolean writeUserToDatabase(TestUser user) {
+        Database db = new Database();
+        //check if user already exists
+        TestUser tmpUser = readUserDataFromDatabase(user.getUsername());
+        //System.out.println("found user: " + tmpUser.toString());
+
+
+        if(user.getUsername().equals(tmpUser.getUsername())){
+            //System.out.println("user already here");
+            return false;
+        }
         System.out.println("  save item: " + user);
-        daoDb.saveUser( new TestUser(
-                user.getUsername(),
-                user.getPassword(),
-                user.getToken()
-        ) );
+            //store new user in db
+            db.saveUser( new TestUser(
+                    user.getUsername(),
+                    user.getPassword(),
+                    user.getToken()
+            ) );
+            return true;
+
     }
 
     /*public static void writeCardsListToDatabase(List<TestCardDB> data) {
@@ -98,20 +110,32 @@ public class DatabaseOperations {
         }
     }
 
-    public static void aquirePackageByToken(String token){
+    public static int aquirePackageByToken(String token){
         Database db = new Database();
-        //get ids of cards from package
-        List<String> ids = db.getIdsOfPackage();
-        System.out.println(ids.toString());
-        //update cards with token
-        for(int i = 0;i < 5;i++){
-            System.out.println("  update item: " + ids.get(i));
-            db.updateCard(ids.get(i), token);
+        //check if enough money left
+        if(db.getCoinsByToken(token) < 5){
+           return -1;
+        }else{
+            //get ids of cards from package
+            List<String> ids = db.getIdsOfPackage();
+            System.out.println(ids.toString());
+            if(ids.isEmpty()){
+                return 1;
+            }else{
+                //update cards with token
+                for(int i = 0;i < 5;i++){
+                    System.out.println("  update item: " + ids.get(i));
+                    db.updateCard(ids.get(i), token);
+                }
+                //get Package ID
+                int packId = db.getIdOfPackage();
+                //delete package
+                db.deletePackage(packId);
+                //update coins of user
+                db.updateCoinsByToken(token);
+                return 0;
+            }
         }
-        //get Package ID
-        int packId = db.getIdOfPackage();
-        //delete package
-        db.deletePackage(packId);
     }
 
     public static List<TestCardDB> readCardsFromDatabaseByToken(String token){
